Fixed the map  hex numbers to behave properly with new translation code. Made them considerably prettier while I was at it.
Also changed a few things around in Thud.java to make it easier to add listeners for the menus. This is prep work for (a) moving menu handling code to a new class and (b) making all windows respond to the  menu commands instead of just the main text window.